Timer now works (displaying loadable sudokus doesnt for some reason, might just be me)

+++ src/timeHelper.h
@@ -16,13 +16,13 @@
 typedef struct stopWatch {
     clock_t startTime;
     clock_t endTime;
-    double timeInSeconds;
+    int timeInSeconds;
 } StopWatch;
 
 StopWatch startTimer();
 
 void stopTimer(StopWatch *watch);
 
-double getTimeInSeconds(StopWatch *watch);
+int getTimeInSeconds(StopWatch *watch);
 
 #endif
